handmade-products/face-recognition/requirements.txt
image2face
opencv-python
numpy
imutils
python-dotenv
Pillow
